Notification feature completed!

+++ INSTA/Instagram/post/views.py
@@ -154,6 +154,18 @@ def showPost(request, slug):
     data["request_username"] = username 
     data["request_img_src"] = img_src
 
+    #Fetching the unseen notifications
+    cmnd = """
+    SELECT COUNT(*)
+    FROM NOTIFICATION
+    WHERE TO_ID = :userid AND IS_SEEN = 0
+    """
+    c = connection.cursor()
+    c.execute(cmnd, [userid])
+    row = c.fetchone()
+    total_unseen = row[0] 
+    data["total_unseen"] = total_unseen
+
 
     connection.close()
     return render(request, 'post/postpage.html',  data)
@@ -219,14 +231,15 @@ def likepost(request):
         c.execute(cmnd, [userid, postid]) 
         connection.commit()
 
-        #insert into notification table
-        cmnd = """
-        INSERT INTO NOTIFICATION(FROM_ID, TO_ID,CONTENT, RELATED_POST_ID)  
-        VALUES(:user_id, :poster_id, :type , :post_id)
-        """
-        c = connection.cursor()
-        c.execute(cmnd, [userid, poster_id,"like", postid]) 
-        connection.commit()
+        if(not(userid==poster_id)):
+            #insert into notification table
+            cmnd = """
+            INSERT INTO NOTIFICATION(FROM_ID, TO_ID,CONTENT, RELATED_POST_ID)  
+            VALUES(:user_id, :poster_id, :type , :post_id)
+            """
+            c = connection.cursor()
+            c.execute(cmnd, [userid, poster_id,"like", postid]) 
+            connection.commit()
 
     else : #if already liked then dislike 
         cmnd = """
@@ -237,15 +250,15 @@ def likepost(request):
         c.execute(cmnd, [userid, postid])
         connection.commit()
 
-
-        #delete from notification table
-        cmnd = """
-        DELETE FROM NOTIFICATION
-        WHERE FROM_ID = :user_id AND TO_ID = :poster_id AND CONTENT = :type AND RELATED_POST_ID = :post_id 
-        """
-        c = connection.cursor()
-        c.execute(cmnd, [userid, poster_id,"like", postid]) 
-        connection.commit()
+        if(not(userid==poster_id)):
+            #delete from notification table
+            cmnd = """
+            DELETE FROM NOTIFICATION
+            WHERE FROM_ID = :user_id AND TO_ID = :poster_id AND CONTENT = :type AND RELATED_POST_ID = :post_id 
+            """
+            c = connection.cursor()
+            c.execute(cmnd, [userid, poster_id,"like", postid]) 
+            connection.commit()
 
     cmnd = """
     SELECT COUNT(*)
@@ -276,6 +289,8 @@ def likepost(request):
 def postComment(request, slug):
     if(request.method == 'POST'):
         comment = request.POST.get('comment')
+        print(slug)
+        print(comment)
 
         dsn_tns = cx_Oracle.makedsn('localhost', '1521', service_name='ORCL')
         connection = cx_Oracle.connect(user='insta', password='insta', dsn=dsn_tns)
@@ -310,8 +325,17 @@ def postComment(request, slug):
         c.execute(cmnd, [commentid, comment])
         connection.commit()
 
-        
+        #get the poster id
         postid = slug
+        cmnd = """
+        SELECT USER_ID
+        FROM POST
+        WHERE POST_ID = :postid
+        """
+        c = connection.cursor()
+        c.execute(cmnd, [postid])
+        row = c.fetchone()
+        poster_id = row[0]
 
         cmnd = """
         INSERT INTO USER_POST_COMMENT(COMMENTER_ID, POST_ID, COMMENT_ID)
@@ -334,8 +358,22 @@ def postComment(request, slug):
         c.execute(cmnd, [commentid, parent_comment_id])
         connection.commit()
 
+        if(not(commenter_id==poster_id)):
+            #insert into notification table
+            cmnd = """
+            INSERT INTO NOTIFICATION(FROM_ID, TO_ID,CONTENT, RELATED_POST_ID)  
+            VALUES(:commenter_id, :poster_id, :type , :post_id)
+            """
+            if(parent_comment_id == commentid):
+                content = "comment"
+            else :
+                content = "reply"
+            c = connection.cursor()
+            c.execute(cmnd, [commenter_id, poster_id,content, postid]) 
+            connection.commit()
 
         connection.close()
+
         return redirect(f"/post/{slug}")
     else :
         return HttpResponse('404-Not found')
